Changes requested in review.
