+++ src/vulkan/pipeline.cc
@@ -279,8 +279,9 @@ Result Pipeline::AddBufferDescriptor(const BufferCommand* cmd) {
   if (cmd->IsPushConstant())
     return AddPushConstantBuffer(cmd->GetBuffer(), cmd->GetOffset());
   if (!cmd->IsSSBO() && !cmd->IsUniform() && !cmd->IsStorageImage() &&
-      !cmd->IsSampledImage() && !cmd->IsCombinedImageSampler())
+      !cmd->IsSampledImage() && !cmd->IsCombinedImageSampler()) {
     return Result("Pipeline::AddBufferDescriptor not supported buffer type");
+  }
 
   Descriptor* desc;
   Result r =
@@ -304,8 +305,8 @@ Result Pipeline::AddBufferDescriptor(const BufferCommand* cmd) {
     } else if (cmd->IsCombinedImageSampler()) {
       auto image_desc = MakeUnique<ImageDescriptor>(
           cmd->GetBuffer(), DescriptorType::kCombinedImageSampler, device_,
-          cmd->GetDescriptorSet(), cmd->GetBinding(),
-          cmd->GetBuffer()->GetSampler());
+          cmd->GetDescriptorSet(), cmd->GetBinding());
+      image_desc->SetAmberSampler(cmd->GetBuffer()->GetSampler());
       descriptors.push_back(std::move(image_desc));
     } else {
       auto desc_type = cmd->IsSSBO() ? DescriptorType::kStorageBuffer
